+++ logs-data/lang/langage.js
@@ -1,15 +1,13 @@
-function lang(value = "fr"){
-
-    if(value === "fr"){
-      const fr = require("./app-dic-fr.json")
-      return fr
-    }else{
-      const en = require("./app-dic-en.json")
-      return en
+function lang(value = 'fr') {
+    if (value === 'fr') {
+        const fr = require('./app-dic-fr.json')
+        return fr
+    } else {
+        const en = require('./app-dic-en.json')
+        return en
     }
-  
-  }
+}
 
-  module.exports = {
-      lang
-  }
+module.exports = {
+    lang,
+}