+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # V.1.5
 
 Change File organisation
+
 # v1.0.1
 
-- change module langage
-- add version systeme
+-   change module langage
+-   add version systeme
+
 # V1.0.0
 
- - Création de l'app 
+-   Création de l'app